Clarify naming a little
