+++ app.js
@@ -37,34 +37,34 @@ app.get('/api/login', auth.tokenToBody, auth.loginToManagebac, (req, res) => {
 app.post('/api/login', upload.none(), auth.loginToManagebac, auth.createSardonyxToken, mb.loadDefaults);
 
 // Load class
-app.get('/api/class/:resourceId/overview', auth.tokenToCookie, mb.loadUrl('classes'), mb.retrieve, mb.loadOverview);
-app.get('/api/class/:resourceId/assignments', auth.tokenToCookie, mb.loadUrl('classes', 'assignments'), mb.retrieve, mb.loadAssignments);
-app.get('/api/class/:resourceId/messages', auth.tokenToCookie, mb.loadUrl('classes', 'discussions'), mb.retrieve, mb.loadMessages);
+app.get('/api/class/:resourceId/overview', auth.tokenToCookie, mb.createUrl('classes'), mb.scrape, mb.loadOverview);
+app.get('/api/class/:resourceId/assignments', auth.tokenToCookie, mb.createUrl('classes', 'assignments'), mb.scrape, mb.loadAssignments);
+app.get('/api/class/:resourceId/messages', auth.tokenToCookie, mb.createUrl('classes', 'discussions'), mb.scrape, mb.loadMessages);
 
 // Load group 
-app.get('/api/group/:resourceId/overview', auth.tokenToCookie, mb.loadUrl('groups'), mb.retrieve, mb.loadOverview);
-app.get('/api/group/:resourceId/messages', auth.tokenToCookie, mb.loadUrl('groups', 'discussions'), mb.retrieve, mb.loadMessages);
+app.get('/api/group/:resourceId/overview', auth.tokenToCookie, mb.createUrl('groups'), mb.scrape, mb.loadOverview);
+app.get('/api/group/:resourceId/messages', auth.tokenToCookie, mb.createUrl('groups', 'discussions'), mb.scrape, mb.loadMessages);
 
 // Load assignment
-app.get('/api/class/:resourceId/assignments/:destinationId', auth.tokenToCookie, mb.loadUrl('classes', 'assignments'), mb.retrieve, mb.loadAssignment);
+app.get('/api/class/:resourceId/assignments/:destinationId', auth.tokenToCookie, mb.createUrl('classes', 'assignments'), mb.scrape, mb.loadAssignment);
 
 // Load message 
-app.get('/api/class/:resourceId/messages/:destinationId', auth.tokenToCookie, mb.loadUrl('classes', 'discussions'), mb.retrieve, mb.loadMessage);
-app.get('/api/group/:resourceId/messages/:destinationId', auth.tokenToCookie, mb.loadUrl('groups', 'discussions'), mb.retrieve, mb.loadMessage);
+app.get('/api/class/:resourceId/messages/:destinationId', auth.tokenToCookie, mb.createUrl('classes', 'discussions'), mb.scrape, mb.loadMessage);
+app.get('/api/group/:resourceId/messages/:destinationId', auth.tokenToCookie, mb.createUrl('groups', 'discussions'), mb.scrape, mb.loadMessage);
 
 // Send message 
-app.post('/api/class/:resourceId/messages', auth.tokenToCookie, mb.loadUrl('classes', 'discussions'), mb.sendMessage);
-app.post('/api/group/:resourceId/messages', auth.tokenToCookie, mb.loadUrl('groups', 'discussions'), mb.sendMessage);
+app.post('/api/class/:resourceId/messages', auth.tokenToCookie, mb.createUrl('classes', 'discussions'), mb.sendMessage);
+app.post('/api/group/:resourceId/messages', auth.tokenToCookie, mb.createUrl('groups', 'discussions'), mb.sendMessage);
 
 // Load notifications 
-app.get('/api/notification', auth.tokenToCookie, mb.loadUrl('notifications'), mb.retrieve, mb.loadNotifications);
-app.get('/api/notification/:resourceId', auth.tokenToCookie, mb.loadUrl('notifications'), mb.retrieve, mb.loadNotification);
+app.get('/api/notification', auth.tokenToCookie, mb.createUrl('notifications'), mb.scrape, mb.loadNotifications);
+app.get('/api/notification/:resourceId', auth.tokenToCookie, mb.createUrl('notifications'), mb.scrape, mb.loadNotification);
 
 // Load CAS
-app.get('/api/cas', auth.tokenToCookie, mb.loadUrl('ib/activity/cas'), mb.retrieve, mb.loadCas);
-app.get('/api/cas/:resourceId', auth.tokenToCookie, mb.loadUrl('ib/activity/cas'), mb.retrieve, mb.loadExperience);
-app.get('/api/cas/:resourceId/answers', auth.tokenToCookie, mb.loadUrl('ib/activity/cas', 'answers'), mb.retrieve, mb.loadAnswers);
-app.get('/api/cas/:resourceId/reflections', auth.tokenToCookie, mb.loadUrl('ib/activity/cas', 'reflections'), mb.retrieve, mb.loadReflections);
+app.get('/api/cas', auth.tokenToCookie, mb.createUrl('ib/activity/cas'), mb.scrape, mb.loadCas);
+app.get('/api/cas/:resourceId', auth.tokenToCookie, mb.createUrl('ib/activity/cas'), mb.scrape, mb.loadExperience);
+app.get('/api/cas/:resourceId/answers', auth.tokenToCookie, mb.createUrl('ib/activity/cas', 'answers'), mb.scrape, mb.loadAnswers);
+app.get('/api/cas/:resourceId/reflections', auth.tokenToCookie, mb.createUrl('ib/activity/cas', 'reflections'), mb.scrape, mb.loadReflections);
 
 module.exports = app;
 // app.js and server.js are split for testing reasons
